+++ polystatistics.h
@@ -72,7 +72,7 @@ typedef struct polystatistics {
 #else
     int psTimers[N_PS_TIMES];
 #endif
-    long psUser[N_PS_USER];
+    intptr_t psUser[N_PS_USER];
 } polystatistics;
 
 #endif // POLY_STATISTICS_INCLUDED
